Separated Functionality From PlayerController
The PlayerController input functions now call LevelManager functions.
This is so the PlayerController isn't connected to more than the LevelManager.
The LevelManager should have control over everything in the scene. Objects that aren't the level manager should have limited access and control over other objects. This creates a flow to the structure.